refactor(config): extract harness posture module
Move the harness posture/profile data model, built-in seeds, and profile matching helpers out of lib.rs into a focused harness module.\n\nRe-export the public harness types and built_in_harness_profiles from the crate root so existing callers and tests keep the same API.\n\nThis is a behavior-preserving config split slice for #3311.
